pyproject.toml
[project]
name = "agent-middleware"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.4",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "pydantic>=2.11.4",
    "pyodbc>=5.2.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]
